requirements.txt
discord
python-dotenv
pandas
fuzzywuzzy
python-Levenshtein